Improve error handling, client configuration, and flush concurrency (#3197)

+++ aws-kinesis/src/main/java/com/amplifyframework/recordcache/RecordClient.kt
@@ -1,52 +1,45 @@
 package com.amplifyframework.recordcache
 
-import kotlin.mapCatching
+import java.util.concurrent.atomic.AtomicBoolean
 
-internal class RecordClient<E : Exception>(
+internal class RecordClient(
     private val sender: RecordSender,
-    private val storage: RecordStorage,
-    private val exceptionMapper: (RecordCacheException) -> E
+    private val storage: RecordStorage
 ) {
-    suspend fun record(record: RecordInput): RecordResult = mapErrorResult(
-        storage.addRecord(record).mapCatching { RecordData() }
-    )
+    private val isFlushing = AtomicBoolean(false)
+    suspend fun record(record: RecordInput): RecordResult = storage.addRecord(record).mapCatching { RecordData() }
 
-    suspend fun flush(): FlushResult = mapErrorResult(
-        runCatching {
-            val r = storage.getRecordsByStream()
-                .getOrThrow()
-                .map { records ->
-                    val streamName = records.first().streamName
-                    val result = sender.putRecords(streamName, records).getOrThrow()
-
-                    val deleteSuccessful = storage.deleteRecords(result.successfulIds)
-                    val deleteFailed = storage.deleteRecords(result.failedIds)
-                    val incrementRetry = storage.incrementRetryCount(result.retryableIds)
-
-                    // Ensure all updates are triggered before checking for exceptions
-                    deleteSuccessful.getOrThrow()
-                    deleteFailed.getOrThrow()
-                    incrementRetry.getOrThrow()
-                    
-                    result.successfulIds
-                }
-                .map { it.size }.sum()
-            FlushData(r)
+    suspend fun flush(): FlushResult {
+        // Guard against concurrent flushes to return early
+        if (!isFlushing.compareAndSet(false, true)) {
+            return Result.success(FlushData(recordsFlushed = 0, flushInProgress = true))
         }
-    )
+        return try {
+            runCatching {
+                val r = storage.getRecordsByStream()
+                    .getOrThrow()
+                    .map { records ->
+                        val streamName = records.first().streamName
+                        val result = sender.putRecords(streamName, records).getOrThrow()
 
-    suspend fun clearCache(): ClearCacheResult = mapErrorResult(storage.clearRecords())
+                        val deleteSuccessful = storage.deleteRecords(result.successfulIds)
+                        val deleteFailed = storage.deleteRecords(result.failedIds)
+                        val incrementRetry = storage.incrementRetryCount(result.retryableIds)
 
-    fun <T> mapErrorResult(result: Result<T>): Result<T> = if (result.isSuccess) {
-        result
-    } else {
-        val exc = result.exceptionOrNull()
-        Result.failure(
-            exceptionMapper(
-                exc as? RecordCacheException
-                    ?: // TODO: What to pass here?
-                    RecordCacheException("", "", exc)
-            )
-        )
+                        // Ensure all updates are triggered before checking for exceptions
+                        deleteSuccessful.getOrThrow()
+                        deleteFailed.getOrThrow()
+                        incrementRetry.getOrThrow()
+
+                        result.successfulIds
+                    }
+                    .map { it.size }.sum()
+                FlushData(r)
+            }
+        } finally {
+            isFlushing.set(false)
+        }
     }
+
+    suspend fun clearCache(): ClearCacheResult = storage.clearRecords()
 }
